refactor(api): move AdminMember model
