+++ api/tests/tooling/domain-builder/factory/build-admin-member.js
@@ -1,4 +1,4 @@
-import { AdminMember } from '../../../../lib/domain/models/AdminMember.js';
+import { AdminMember } from '../../../../src/shared/domain/models/AdminMember.js';
 import { PIX_ADMIN } from '../../../../src/authorization/domain/constants.js';
 
 const { ROLES } = PIX_ADMIN;
